+++ node/tests/listSpecs.go
@@ -18,7 +18,7 @@ func SpecTestListString(t *testing.T, np ipld.NodePrototype) {
 			la.AssembleValue().AssignString("three")
 		})
 		t.Run("reads back out", func(t *testing.T) {
-			Wish(t, n.Length(), ShouldEqual, 3)
+			Wish(t, n.Length(), ShouldEqual, int64(3))
 
 			v, err := n.LookupByIndex(0)
 			Wish(t, err, ShouldEqual, nil)
@@ -38,25 +38,25 @@ func SpecTestListString(t *testing.T, np ipld.NodePrototype) {
 			Wish(t, itr.Done(), ShouldEqual, false)
 			idx, v, err := itr.Next()
 			Wish(t, err, ShouldEqual, nil)
-			Wish(t, idx, ShouldEqual, 0)
+			Wish(t, idx, ShouldEqual, int64(0))
 			Wish(t, must.String(v), ShouldEqual, "one")
 
 			Wish(t, itr.Done(), ShouldEqual, false)
 			idx, v, err = itr.Next()
 			Wish(t, err, ShouldEqual, nil)
-			Wish(t, idx, ShouldEqual, 1)
+			Wish(t, idx, ShouldEqual, int64(1))
 			Wish(t, must.String(v), ShouldEqual, "two")
 
 			Wish(t, itr.Done(), ShouldEqual, false)
 			idx, v, err = itr.Next()
 			Wish(t, err, ShouldEqual, nil)
-			Wish(t, idx, ShouldEqual, 2)
+			Wish(t, idx, ShouldEqual, int64(2))
 			Wish(t, must.String(v), ShouldEqual, "three")
 
 			Wish(t, itr.Done(), ShouldEqual, true)
 			idx, v, err = itr.Next()
 			Wish(t, err, ShouldEqual, ipld.ErrIteratorOverread{})
-			Wish(t, idx, ShouldEqual, -1)
+			Wish(t, idx, ShouldEqual, int64(-1))
 			Wish(t, v, ShouldEqual, nil)
 		})
 	})
